pyproject.toml
[project]
name = "hack-detector"
version = "0.4.4"
description = "DeFi hack incident auto-detection pipeline"
requires-python = ">=3.12"
dependencies = [
    "telethon>=1.37",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "supabase>=2.0",
    "httpx>=0.27",
    "openai>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.7",
]

